added readme for bms and mcu. Added bms_test.hpp. Also fixed mcu parsing and test data

+++ src/mcu.h
@@ -18,7 +18,7 @@ typedef struct ControllerStatus
 
 typedef struct MotorData
 {
-    float rpm;
+    uint16_t rpm;
     float mcuCurrent;
     float mcuVoltage;
     bool mcuFaults[16];
